+++ yash-syntax/src/parser/and_or.rs
@@ -132,6 +132,6 @@ mod tests {
         assert_eq!(*e.location.code.value.borrow(), "foo &&");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 7);
+        assert_eq!(e.location.index.get(), 7);
     }
 }