+++ yash-syntax/src/parser/case.rs
@@ -318,7 +318,7 @@ mod tests {
         assert_eq!(*e.location.code.value.borrow(), ")");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 1);
+        assert_eq!(e.location.index.get(), 1);
     }
 
     #[test]
@@ -332,7 +332,7 @@ mod tests {
         assert_eq!(*e.location.code.value.borrow(), "(esac)");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 2);
+        assert_eq!(e.location.index.get(), 2);
     }
 
     #[test]
@@ -346,7 +346,7 @@ mod tests {
         assert_eq!(*e.location.code.value.borrow(), "(&");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 2);
+        assert_eq!(e.location.index.get(), 2);
     }
 
     #[test]
@@ -360,7 +360,7 @@ mod tests {
         assert_eq!(*e.location.code.value.borrow(), "(foo| |");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 7);
+        assert_eq!(e.location.index.get(), 7);
     }
 
     #[test]
@@ -377,7 +377,7 @@ mod tests {
         assert_eq!(*e.location.code.value.borrow(), "(foo bar");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 6);
+        assert_eq!(e.location.index.get(), 6);
     }
 
     #[test]
@@ -571,7 +571,7 @@ mod tests {
         assert_eq!(*e.location.code.value.borrow(), " case  ");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 8);
+        assert_eq!(e.location.index.get(), 8);
     }
 
     #[test]
@@ -585,7 +585,7 @@ mod tests {
         assert_eq!(*e.location.code.value.borrow(), " case ; ");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 7);
+        assert_eq!(e.location.index.get(), 7);
     }
 
     #[test]
@@ -600,12 +600,12 @@ mod tests {
             assert_eq!(*opening_location.code.value.borrow(), " case x esac");
             assert_eq!(opening_location.code.start_line_number.get(), 1);
             assert_eq!(opening_location.code.source, Source::Unknown);
-            assert_eq!(opening_location.column.get(), 2);
+            assert_eq!(opening_location.index.get(), 2);
         });
         assert_eq!(*e.location.code.value.borrow(), " case x esac");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 9);
+        assert_eq!(e.location.index.get(), 9);
     }
 
     #[test]
@@ -620,11 +620,11 @@ mod tests {
             assert_eq!(*opening_location.code.value.borrow(), "case x in a) }");
             assert_eq!(opening_location.code.start_line_number.get(), 1);
             assert_eq!(opening_location.code.source, Source::Unknown);
-            assert_eq!(opening_location.column.get(), 1);
+            assert_eq!(opening_location.index.get(), 1);
         });
         assert_eq!(*e.location.code.value.borrow(), "case x in a) }");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 14);
+        assert_eq!(e.location.index.get(), 14);
     }
 }