+++ yash-syntax/src/parser/function.rs
@@ -165,7 +165,7 @@ mod tests {
         assert_eq!(*e.location.code.value.borrow(), "( ");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 3);
+        assert_eq!(e.location.index.get(), 3);
     }
 
     #[test]
@@ -187,7 +187,7 @@ mod tests {
         assert_eq!(*e.location.code.value.borrow(), "( ) ");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 5);
+        assert_eq!(e.location.index.get(), 5);
     }
 
     #[test]
@@ -209,7 +209,7 @@ mod tests {
         assert_eq!(*e.location.code.value.borrow(), "() foo ; ");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 4);
+        assert_eq!(e.location.index.get(), 4);
     }
 
     #[test]
@@ -326,6 +326,6 @@ mod tests {
         assert_eq!(*e.location.code.value.borrow(), "()b");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 3);
+        assert_eq!(e.location.index.get(), 3);
     }
 }