+++ yash-syntax/src/parser/simple_command.rs
@@ -236,12 +236,12 @@ mod tests {
             assert_eq!(*opening_location.code.value.borrow(), "(a b");
             assert_eq!(opening_location.code.start_line_number.get(), 1);
             assert_eq!(opening_location.code.source, Source::Unknown);
-            assert_eq!(opening_location.column.get(), 1);
+            assert_eq!(opening_location.index.get(), 1);
         });
         assert_eq!(*e.location.code.value.borrow(), "(a b");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 5);
+        assert_eq!(e.location.index.get(), 5);
     }
 
     #[test]
@@ -255,12 +255,12 @@ mod tests {
             assert_eq!(*opening_location.code.value.borrow(), "(a;b)");
             assert_eq!(opening_location.code.start_line_number.get(), 1);
             assert_eq!(opening_location.code.source, Source::Unknown);
-            assert_eq!(opening_location.column.get(), 1);
+            assert_eq!(opening_location.index.get(), 1);
         });
         assert_eq!(*e.location.code.value.borrow(), "(a;b)");
         assert_eq!(e.location.code.start_line_number.get(), 1);
         assert_eq!(e.location.code.source, Source::Unknown);
-        assert_eq!(e.location.column.get(), 3);
+        assert_eq!(e.location.index.get(), 3);
     }
 
     #[test]
@@ -298,7 +298,7 @@ mod tests {
         assert_eq!(*sc.assigns[0].location.code.value.borrow(), "my=assignment");
         assert_eq!(sc.assigns[0].location.code.start_line_number.get(), 1);
         assert_eq!(sc.assigns[0].location.code.source, Source::Unknown);
-        assert_eq!(sc.assigns[0].location.column.get(), 1);
+        assert_eq!(sc.assigns[0].location.index.get(), 1);
     }
 
     #[test]
@@ -316,19 +316,19 @@ mod tests {
         assert_eq!(*sc.assigns[0].location.code.value.borrow(), "a= b=! c=X");
         assert_eq!(sc.assigns[0].location.code.start_line_number.get(), 1);
         assert_eq!(sc.assigns[0].location.code.source, Source::Unknown);
-        assert_eq!(sc.assigns[0].location.column.get(), 1);
+        assert_eq!(sc.assigns[0].location.index.get(), 1);
         assert_eq!(sc.assigns[1].name, "b");
         assert_eq!(sc.assigns[1].value.to_string(), "!");
         assert_eq!(*sc.assigns[1].location.code.value.borrow(), "a= b=! c=X");
         assert_eq!(sc.assigns[1].location.code.start_line_number.get(), 1);
         assert_eq!(sc.assigns[1].location.code.source, Source::Unknown);
-        assert_eq!(sc.assigns[1].location.column.get(), 4);
+        assert_eq!(sc.assigns[1].location.index.get(), 4);
         assert_eq!(sc.assigns[2].name, "c");
         assert_eq!(sc.assigns[2].value.to_string(), "X");
         assert_eq!(*sc.assigns[2].location.code.value.borrow(), "a= b=! c=X");
         assert_eq!(sc.assigns[2].location.code.start_line_number.get(), 1);
         assert_eq!(sc.assigns[2].location.code.source, Source::Unknown);
-        assert_eq!(sc.assigns[2].location.column.get(), 8);
+        assert_eq!(sc.assigns[2].location.index.get(), 8);
     }
 
     #[test]
@@ -559,7 +559,7 @@ mod tests {
         assert_eq!(*sc.assigns[0].location.code.value.borrow(), "a= ()");
         assert_eq!(sc.assigns[0].location.code.start_line_number.get(), 1);
         assert_eq!(sc.assigns[0].location.code.source, Source::Unknown);
-        assert_eq!(sc.assigns[0].location.column.get(), 1);
+        assert_eq!(sc.assigns[0].location.index.get(), 1);
 
         let next = block_on(parser.peek_token()).unwrap();
         assert_eq!(next.id, Operator(OpenParen));
@@ -580,7 +580,7 @@ mod tests {
         assert_eq!(*sc.assigns[0].location.code.value.borrow(), "a=b()");
         assert_eq!(sc.assigns[0].location.code.start_line_number.get(), 1);
         assert_eq!(sc.assigns[0].location.code.source, Source::Unknown);
-        assert_eq!(sc.assigns[0].location.column.get(), 1);
+        assert_eq!(sc.assigns[0].location.index.get(), 1);
 
         let next = block_on(parser.peek_token()).unwrap();
         assert_eq!(next.id, Operator(OpenParen));